requirements.txt
flask>=3.1.0
python-dotenv>=1.1.0

# Embeddings
sentence-transformers>=4.1.0
torch>=2.1.0

# Vector DB
pinecone>=7.0.0

# LLM
groq>=0.37.1

# Utils
pypdf>=5.6.1
